Cleanup Angular integration
